+++ src/components/TreeList/TreeList.tsx
@@ -28,7 +28,7 @@ export const TreeList = <T,>({
     renderItem: propsRenderItem,
     renderContainer = ListContainer,
     onItemClick: propsOnItemClick,
-    mapItemDataToProps,
+    mapItemDataToContentProps,
 }: TreeListProps<T>) => {
     const uniqId = useUniqId();
     const treeListId = id ?? uniqId;
@@ -71,13 +71,14 @@ export const TreeList = <T,>({
             id: itemId,
             size,
             multiple,
-            mapItemDataToProps,
+            mapItemDataToContentProps,
             onItemClick,
             list,
         });
 
         if (propsRenderItem) {
             return propsRenderItem({
+                id: itemId,
                 data: renderState.data,
                 props: renderState.props,
                 context: renderState.context,