+++ src/components/TreeSelect/types.ts
@@ -15,6 +15,8 @@ import type {UseListParsedStateProps} from '../useList/hooks/useListParsedState'
 export type TreeSelectRenderControlProps<T> = {
     list: UseListResult<T>;
     open: boolean;
+    disabled?: boolean;
+    placeholder?: string;
     toggleOpen(): void;
     clearValue(): void;
     ref: React.Ref<HTMLButtonElement>;
@@ -44,6 +46,8 @@ export interface TreeSelectProps<T, P extends {} = {}>
      */
     title?: string;
     value?: ListItemId[];
+    disabled?: boolean;
+    placeholder?: string;
     defaultValue?: ListItemId[] | undefined;
     popupClassName?: string;
     popupWidth?: SelectPopupProps['width'];